LayoutTests/http/tests/media/modern-media-controls/time-labels-support/long-time.html

<script src="../../../../../resources/js-test-pre.js"></script>
<script src="/media-resources/modern-media-controls/resources/media-controls-loader.js"></script>
<body>
<video src="../../resources/hls/generate-vod.php?duration=8000" style="width: 320px; height: 240px;"></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the display of time labels with a video that lasts an hour or more.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

media.addEventListener("durationchange", function() {
    media.currentTime = 5;
    shouldBeTrue("mediaController.controls.timeControl.useSixDigitsForTimeLabels");
    shouldBecomeEqualToString("mediaController.controls.timeControl.elapsedTimeLabel.element.textContent", "00:00:05", () => {
        debug("");
        shadowRoot.host.remove();
        media.remove();
        finishJSTest();
    });
});

</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
