LayoutTests/http/tests/media/modern-media-controls/time-labels-support/long-time-expected.txt

Testing the display of time labels with a video that lasts an hour or more.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.timeControl.useSixDigitsForTimeLabels is true
PASS mediaController.controls.timeControl.elapsedTimeLabel.element.textContent became "00:00:05"

PASS successfullyParsed is true

TEST COMPLETE
